+++ tests/decode/test3_simple/src/main.c
@@ -432,7 +432,7 @@ uint8_t crash_vector18[] = {
 void test_pet(void)
 {
 	struct Pet pet;
-	uint32_t decode_len;
+	size_t decode_len;
 	uint8_t input[] = {
 		0x83, 0x82, 0x63, 0x66, 0x6f, 0x6f, 0x63, 0x62, 0x61, 0x72,
 		0x48, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08,
@@ -482,7 +482,7 @@ void test_pet_regression(void)
 void test_serial1(void)
 {
 	struct Upload upload;
-	uint32_t decode_len;
+	size_t decode_len;
 	bool ret = cbor_decode_Upload(serial_rec_input1,
 			sizeof(serial_rec_input1), &upload, &decode_len);
 	zassert_true(ret, "decoding failed.");
@@ -505,7 +505,7 @@ void test_serial1(void)
 void test_serial2(void)
 {
 	struct Upload upload;
-	uint32_t decode_len;
+	size_t decode_len;
 	bool ret = cbor_decode_Upload(serial_rec_input2,
 			sizeof(serial_rec_input2), &upload, &decode_len);
 	zassert_true(ret, "decoding failed.");
